backend-fastapi/requirements.txt
fastapi
uvicorn
python-dotenv
openai
langchain
langchain-community
faiss-cpu
httpx==0.25.2
python-multipart==0.0.6